feat(ng/cropper): split Cropper into element + directive + modal + service
Rearchitect the Cropper package to mirror React's three-layer shape:

- MznCropper directive (shell): [mznCropper] attribute, host class
  `mzn-cropper mzn-cropper--{size}`, parity with React <Cropper>.
- MznCropperElement (new, renamed from the old MznCropper): the
  canvas-bearing component. Strips the Angular-only over-implementation
  (resize handles, move/resize logic, 3x3 grid, disabled/min|maxScale
  inputs, public scale/cropArea signals) to match React's fixed crop
  area model. Uses MznSlider + MznTypography internally to mirror
  React's <Slider prefixIcon={MinusIcon} suffixIcon={PlusIcon}> and
  <Typography variant="label-secondary" color="text-fixed-light">.
- MznCropperModal (new): wraps MznModal + MznCropperElement; emits
  cropChange, confirmed (with {canvas, cropArea, imageSrc}),
  cancelled, closed.
- MznCropperModalService.open(options): imperative Promise API using
  createComponent + ApplicationRef to dynamically mount the modal at
  document.body, matching React's createNotifier flow.

Also fix two rendering bugs in MznCropperElement.drawCanvas:
1. Lock canvas CSS width/height before writing width/height attrs so
   the aspect-ratio presentational hint cannot feed back into a
   `width: max-content` ancestor and blow up the modal size on every
   redraw ("整個 modal 都放大" bug).
2. Bind the wheel listener on the canvas element itself, matching
   React Cropper's canvas.addEventListener('wheel', ...) placement.
