+++ packages/ng/cropper/index.ts
@@ -1,5 +1,16 @@
-export { MznCropper } from './cropper.component';
-export type { CropArea } from './cropper.component';
+export { MznCropper } from './cropper.directive';
+export { MznCropperElement } from './cropper-element.component';
+export type { CropArea } from './cropper-element.component';
+export { MznCropperModal } from './cropper-modal.component';
+export type {
+  CropperModalConfirmContext,
+  CropperPropsBase,
+} from './cropper-modal.component';
+export { MznCropperModalService } from './cropper-modal.service';
+export type {
+  CropperModalOpenOptions,
+  CropperModalResult,
+} from './cropper-modal.service';
 export {
   cropToBlob,
   cropToDataURL,
@@ -7,3 +18,4 @@ export {
   loadImage,
 } from './cropper-tools';
 export type { CropToBlobOptions } from './cropper-tools';
+export type { CropperSize } from '@mezzanine-ui/core/cropper';
